% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentTS.2catsignal.R
\name{segmentTS.2catsignal}
\alias{segmentTS.2catsignal}
\title{Categorize Signals}
\usage{
segmentTS.2catsignal(dat, lolim = -999)
}
\arguments{
\item{dat}{data.frame object with variables of data value; variables derived from segmentTS.1matchsignal.}

\item{lolim}{lower limit for consideration of matching; set to low value so that all values potentially match.}
}
\value{
data.frame object with variable 'pos', categorized as above; used in segmentTS.2eqsignal,
}
\description{
This function takes in the time-series data for observed and simulated from segmentTS.1matchsignal.
Categorizes portions of the curve into clear signals {trough, up, no-event, down, or peak}
based on a difference equation to determine first and second derivatives.
}
