% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_functions.R
\name{segmentTS.4segdist}
\alias{segmentTS.4segdist}
\title{Segment Distance}
\usage{
segmentTS.4segdist(obs.seg, sim.seg)
}
\arguments{
\item{obs.seg}{data.frame object with variables derived from segmentTS.3eqsignal}

\item{sim.seg}{data.frame object, variables as in obs.seg, but for simulated data}
}
\value{
list object with 4 outputs: time-series of the matching times and values (poly_t,poly) and the distance statistics (dist_tdiff,dist_vdiff)
}
\description{
This function takes in the time-series data for observed and simulated from segmentTS.3eqsignal.
Matches similar signals in the two time-series. Distance statistics are simulation - observation.
}
