% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentTS.4statsplots.R
\name{segmentTS.4statsplots}
\alias{segmentTS.4statsplots}
\title{Segment-based Statistics & Figures}
\usage{
segmentTS.4statsplots(obs.evnt, sim.evnt, ls.evnt.pos, obs.name = "obs",
  sim.name = "sim", time.units = "days", val.units = NULL,
  save.plot = FALSE, outDir = getwd(), region.name = NULL)
}
\arguments{
\item{obs.evnt}{data.frame object for full time-series of observational data, from segmentTS.1matchsignal.}

\item{sim.evnt}{data.frame object for full time-series of simulated data, from segmentTS.1matchsignal.}

\item{ls.evnt.pos}{list object from segmentTS.3eqsignal with positions of
the peaks and troughs in the full time-series. Defines the boundaries of 
the segments.}

\item{obs.name}{name of observational data for table (string).}

\item{sim.name}{name of simulated data for table (string).}

\item{time.units}{units of time for calculating period length in the time-series; default is days.}

\item{val.units}{units of the value for the variable in the time-series.}

\item{save.plot}{save plot as a pdf (TRUE/FALSE); default is FALSE; default out is the current working directory.}

\item{outDir}{location to save plot; default location is getwd()).}

\item{region.name}{name of underlying region for table (string).
Default is 'null_region'; not important unless evaluating multiple regions.
At minimum, requires variables of values and date (YYYY-MM-DD)}
}
\value{
a data.frame object segment-based statistical summaries for all segments, obs and sim.
}
\description{
This function collates segment-based statistics based on the 
observed and simulated time-series.
}
