% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_functions.R
\name{segmentTS.catsignal}
\alias{segmentTS.catsignal}
\title{Categorize Signals}
\usage{
segmentTS.catsignal(dat = df, lolim = -999)
}
\arguments{
\item{dat}{data.frame object with variables of data value; variables derived from SeriesDist.1X}
}
\value{
data.frame object with variable 'pos', categorized as above; used in SeriesDist.3X
}
\description{
This function takes in the time-series data for observed and simulated from segmentTS.matchsignal.
Categorizes portions of the curve into clear signals {trough, up, no-event, down, or peak}
based on a difference equation to determine first and second derivatives.
}
