% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_functions.R
\name{segmentTS.eqsignal}
\alias{segmentTS.eqsignal}
\title{Equalize Signals}
\usage{
segmentTS.eqsignal(obs.evnt = df, sim.evnt = df, val.mindays = 250,
  manual_removal = NULL)
}
\arguments{
\item{obs.evnt}{data.frame object with variables derived from SeriesDist.3X}

\item{sim.evnt}{data.frame, variables as in obs.evnt, but for simulated data.}

\item{val.mindays}{integer number of timesteps (days) between peaks troughs; helps to remove false peaks and troughs.}

\item{manual_removal}{section of code identifying which peak or trough to remove from the time-series.
By visual inspection, if the first peak/trough is a false peak/trough,
then pass code as obs.peak[c(-1),] or obs.trough[c(-1),]. Add multiple removals via obs.peak[c(-1,-2,...),].
Pass code for multiple signals with semi-colons as in manual_removal="obs.peak[c(-1),]; sim.peak[c(-1),]"}
}
\value{
list object with two data.frames, with number of peaks,troughs equalized; used in SeriesDist.4X
}
\description{
This function takes in the time-series data for observed and simulated from segmentTS.catsignal.
Attempts to equalize the number of signals in simulated time-series to match number of signals in the obs.
Modify the criteria for removing false peaks/troughs in this section. 
i.e., remove short-term signals and focus on seasonal patterns of rise and fall.
Also, set individual criteria to smooth signal characterization for different simulated time-series.
}
