% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentTS.mkdf.R
\name{segmentTS.mkdf}
\alias{segmentTS.mkdf}
\title{Make data frames for event/signal type of time-series data}
\usage{
segmentTS.mkdf(df.obs, df.sim, func.var, date.var)
}
\arguments{
\item{df.obs}{data.frame object based on ccgcrv output for observational data.
At minimum, requires variables of values and date (YYYY-MM-DD)}

\item{df.sim}{data.frame object based on ccgcrv output for simulated data
At minimum, requires variables of values and date (YYYY-MM-DD)}

\item{func.var}{name of variable for data values. Must have same name in both df.obs and df.sim.}

\item{date.var}{name of variable for dates. Must have same name in both df.obs and df.sim.}
}
\value{
a list object with two data.frames, df.obs.evnt, df.sim.evnt formatted with new variable names
}
\description{
This function makes the empty data frames for storing time-series event/signal data.
Removes data in simulated dataset that does not overlap (in time) with observed data.
}
