% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_functions.R
\name{segmentTS.segdist}
\alias{segmentTS.segdist}
\title{Segment Distance}
\usage{
segmentTS.segdist(obs.seg, sim.seg)
}
\arguments{
\item{obs.evnt}{data.frame object with variables derived from SeriesDist.3X}

\item{sim.evnt}{data.frame, variables as in obs.evnt, but for simulated data}
}
\value{
list object with 4 outputs: time-series of the matching times and values (poly_t,poly) and the distance statistics (dist_tdiff,dist_vdiff)
}
\description{
This function takes in the time-series data for observed and simulated from segmentTS.eqsignal.
Matches similar signals in the two time-series. Distance statistics are simulation - observation.
}
