% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_functions.R
\name{SeriesDist.1MatchSignals}
\alias{SeriesDist.1MatchSignals}
\title{Match Signals}
\usage{
SeriesDist.1MatchSignals(obs.evnt = df, sim.evnt = df, limit4match = 0)
}
\arguments{
\item{obs.evnt}{data.frame object with variables of start time (decimal.date), end time (decimal.date), match (integer)}

\item{sim.evnt}{data.frame, as above, but for simulated data}

\item{limit4match}{number to search the vector in obs.evnt,sim.evnt for similar events (set to zero)}
}
\value{
a list with both obs.evnt,sim.evnt with an additional variable for the matching index; used in SeriesDist.2X
}
\description{
This function matches the number of events in the time series. It is based
on Ehret and Zehe's (2011) conception of event-type signals. But for our purposes,
we treat the full time series as a single event and break up the time series into segements.
..this function is generally not used as we set all data points as the same 'event',
..but we keep this function here for future application.
}
