% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_functions.R
\name{SeriesDist.2CatSignals}
\alias{SeriesDist.2CatSignals}
\title{Categorize Signals}
\usage{
SeriesDist.2CatSignals(dat = df, lolim = -999)
}
\arguments{
\item{dat}{data.frame object with variables of data value; variables derived from SeriesDist.1X}
}
\value{
data.frame object with variable 'pos', categorized as above; used in SeriesDist.3X
}
\description{
This function takes in the time-series data for observed and simulated from SeriesDist.1X.
Categorizes portions of the curve into clear signals of trough, up, no-event, down, or peak,
..based on a difference equation to determine first and second derivatives
}
