% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segment_functions.R
\name{SeriesDist.3EqualizeSignals}
\alias{SeriesDist.3EqualizeSignals}
\title{Equalize Signals}
\usage{
SeriesDist.3EqualizeSignals(obs.evnt = df, sim.evnt = df, region_plot,
  is.dgvm = "FALSE")
}
\arguments{
\item{obs.evnt}{data.frame object with variables derived from SeriesDist.3X}

\item{sim.evnt}{data.frame, variables as in obs.evnt, but for simulated data}
}
\value{
list object with two data.frames, with number of peaks,troughs equalized; used in SeriesDist.4X
}
\description{
This function takes in the time-series data for observed and simulated from SeriesDist.2X.
Attempts to equalize the number of signals in simulated time-series to match number of signals in the obs.
Modify the criteria for removing false peaks/troughs in this section. 
..i.e., remove short-term signals and focus on seasonal patterns of rise and fall
..also, set individual criteria to smooth signal characterization for different simulated time-seres
}
